# Conflicts:
#	CMakeLists.txt
#	src/brace_cleanup.cpp
#	src/chunk_list.cpp
#	src/indent.cpp
#	src/output.cpp
#	src/parse_frame.cpp
#	src/uncrustify.cpp
#	src/universalindentgui.cpp

+++ src/chunk_list.cpp
@@ -502,7 +502,7 @@ static void chunk_log(chunk_t* pc, const char* text)
       else if (is_valid(next))   { chunk_log_msg(next, log, " @ before" ); }
       else if (is_valid(prev))   { chunk_log_msg(prev, log, " @ after"  ); }
 
-      LOG_FMT(log, " stage=%d", cpd.unc_stage);
+      LOG_FMT(log, " stage=%d", (int)cpd.unc_stage);
       log_func_stack_inline(log);
    }
 }
